docs: move sidebar.mjs to doc/site/ (site config, not content)
sidebar.mjs is Astro/Starlight configuration — it belongs alongside
astro.config.mjs in doc/site/, not in the content directory doc/md/.
The consumer site reads it from doc/site/sidebar.mjs via the pinned
submodule; the rsync that copies doc/md/ to the docs directory never
touches doc/site/ so no exclusion is needed.
